Summary_department

Cette fonction génère un résumé statistique des élus d'un département, y compris les informations sur les âges, les élus les plus âgés et les plus jeunes, ainsi que les communes ayant les moyennes d'âge les plus faibles et les plus élevées.

Usage

Summary_departement(x)

Arguments

x

Un data.frame contenant les informations des élus. Il doit avoir les colonnes suivantes :

"Code.du.département" : Le code du département.

"Libellé.du.département" : Le nom du département.

"Libellé.de.la.commune" : Le nom de la commune.

"Nom.de.l.élu" : Le nom de l'élu.

"Prénom.de.l.élu" : Le prénom de l'élu.

"Date.de.naissance" : La date de naissance de l'élu (au format "jour/mois/année").

Value

Aucune valeur n'est retournée, mais la fonction affiche les résultats suivants :

Le nom du département.

Le nombre de communes dans le département.

Le nombre d'élu.e.s dans le département.

La distribution des âges des élu.e.s.

Le nom, le prénom, l'âge et la commune de l'élu.e le plus âgé(e).

Le nom, le prénom, l'âge et la commune de l'élu.e le plus jeune.

Le nom de la commune ayant la moyenne d'âge la plus faible, ainsi que la distribution des âges des élu.e.s de cette commune.

Le nom de la commune ayant la moyenne d'âge la plus élevée, ainsi que la distribution des âges des élu.e.s de cette commune.

Examples

# Exemple d'utilisation
# Supposons que df est un dataframe contenant les données nécessaires
df <- data.frame(
  `Code.du.département` = c("01", "01", "01"),
  `Libellé.du.département` = c("Ain", "Ain", "Ain"),
  `Libellé.de.la.commune` = c("Bourg-en-Bresse", "Ambérieu-en-Bugey", "Lagnieu"),
  `Nom.de.l.élu` = c("Dupont", "Martin", "Bernard"),
  `Prénom.de.l.élu` = c("Pierre", "Jacques", "Lucie"),
  `Date.de.naissance` = c("01/01/1970", "15/03/1980", "23/06/1990")
)

# Appel de la fonction avec le dataframe exemple
Summary_departement(df)
#> Nom du département: Ain 
#> Nombre de communes: 3 
#> Nombre d'élu.e.s: 3 
#> Distribution des âges des élu.e.s:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>      35      40      45      45      50      55 
#> 
#> L'élu.e le/la plus âgé(e): Dupont Pierre - Age: 55 - Commune: Bourg-en-Bresse 
#> 
#> L'élu.e le/la plus jeune: Bernard Lucie - Age: 35 - Commune: Lagnieu 
#> 
#> Commune avec la moyenne d'âge la plus faible: Lagnieu 
#> Distribution des âges des élu.e.s de cette commune:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>      35      35      35      35      35      35 
#> 
#> Commune avec la moyenne d'âge la plus élevée: Bourg-en-Bresse 
#> Distribution des âges des élu.e.s de cette commune:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>      55      55      55      55      55      55