Générer un rapport basé sur les élections françaises de 2024

Cette fonction génère un rapport en utilisant un fichier Quarto (.qmd) pour analyser les caractéristiques des élus lors des élections françaises de 2024, en se basant sur les paramètres 'code_commune' et 'code_departement'. Le rapport sera rendu dans le fichier spécifié par l'utilisateur ou choisi via une boîte de dialogue si le chemin de sortie n'est pas précisé.

Usage

generer_rapport(code_commune, code_departement, output = NULL)

Arguments

code_commune

Un numéro de commune (doit être un nombre à 4 chiffres minimum).

code_departement

Un numéro de département (doit être un nombre).

output

Le chemin d'enregistrement du rapport généré. Si non spécifié, une boîte de dialogue permettra à l'utilisateur de choisir l'emplacement du fichier.

Value

Cette fonction génère un fichier rapport en format HTML ou un autre format spécifié dans le fichier Quarto.

Examples

# Exemple d'utilisation de la fonction
# Générer un rapport pour la commune 75056 et le département 75
generer_rapport(1001, 1, output = "chemin/vers/le/rapport.html")
#> Vous avez choisi le numéro de commune: 1001
#> Vous avez choisi le numéro de département: 1
#> Error in setwd(output_dir): cannot change working directory

# Si l'utilisateur ne spécifie pas le chemin de sortie, une boîte de dialogue sera ouverte pour sélectionner l'emplacement du fichier
generer_rapport(1001, 1)
#> Vous avez choisi le numéro de commune: 1001
#> Vous avez choisi le numéro de département: 1
#> 
#> 
#> processing file: rapport.qmd
#> 1/33                   
#> 2/33 [unnamed-chunk-1] 
#> 3/33                   
#> 4/33 [unnamed-chunk-2] 
#> 5/33                   
#> 6/33 [unnamed-chunk-3] 
#> 7/33                   
#> 8/33 [unnamed-chunk-4] 
#> 9/33                   
#> 10/33 [unnamed-chunk-5] 
#> 11/33                   
#> 12/33 [unnamed-chunk-6] 
#> 13/33                   
#> 14/33 [unnamed-chunk-7] 
#> 15/33                   
#> 16/33 [unnamed-chunk-8] 
#> 17/33                   
#> 18/33 [unnamed-chunk-9] 
#> 19/33                   
#> 20/33 [unnamed-chunk-10]
#> 21/33                   
#> 22/33 [unnamed-chunk-11]
#> 23/33                   
#> 24/33 [unnamed-chunk-12]
#> 25/33                   
#> 26/33 [unnamed-chunk-13]
#> 27/33                   
#> 28/33 [unnamed-chunk-14]
#> 29/33                   
#> 30/33 [unnamed-chunk-15]
#> 31/33                   
#> 32/33 [unnamed-chunk-16]
#> 33/33                   
#> output file: rapport.knit.md
#> 
#> pandoc --output ..\..\..\..\..\..\OneDrive\Escritorio\unicornio\rapport1.html
#>   to: html
#>   standalone: true
#>   section-divs: true
#>   html-math-method: mathjax
#>   wrap: none
#>   default-image-extension: png
#>   
#> metadata
#>   document-css: false
#>   link-citations: true
#>   date-format: long
#>   lang: en
#>   title: rapport
#>   author: Àlex Chafes Martorell
#>   
#> Output created: ..\..\..\..\..\..\OneDrive\Escritorio\unicornio\rapport1.html
#> 
#> 
#> Rapport généré avec succès à: C:\Users\aalex\OneDrive\Escritorio\unicornio\rapport1.html